Moved the localStorage file to the lib directory.
