+++ src/lib/localStorage.js
@@ -15,6 +15,6 @@ export const saveState = (state) => {
     const serializedState = JSON.stringify(state);
     localStorage.setItem('state', serializedState);
   } catch (err) {
-    // ignore write errors
+    // ignore write errors for now
   }
 };
